gtm/requirements.txt
docker==4.2.0
GitPython==2.1.10
gitdb2==2.0.6
PyYAML==5.3.1
